% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{filterByMeta}
\alias{filterByMeta}
\title{Filter CKAN by Meta}
\usage{
filterByMeta(datList, meta = "")
}
\arguments{
\item{datList}{(list) output from the Pandora API}

\item{meta}{(character) string for filtering all meta information}
}
\value{
(list) a datList where the entries meta data contains the string 'meta'
}
\description{
Filter CKAN by Meta
}
