% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{getFileTypes}
\alias{getFileTypes}
\title{Get File Types}
\usage{
getFileTypes(
  repository = "",
  network = "",
  pattern = "",
  order = TRUE,
  packageList = data.frame()
)
}
\arguments{
\item{repository}{(character) name of Pandora repository}

\item{network}{(character) name of Pandora network}

\item{pattern}{(character) string for meta information search}

\item{order}{(logical) if TRUE order dataframe alphabetically by name}

\item{packageList}{(data.frame) optional, output of callAPI() from a previous call to the
Pandora API.}
}
\value{
(data.frame) containing available file types within a repository
}
\description{
Get all available file types of a repository or those within a
specific network or within a specific repository
optional filtering of meta information for a given string
}
