% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{getNetworks}
\alias{getNetworks}
\title{Get Networks}
\usage{
getNetworks(pattern = "", order = TRUE)
}
\arguments{
\item{pattern}{(character) string for meta information search}
}
\value{
a data.frame giving the "name" and "display_name" of available Pandora networks
}
\description{
Formerly getCKANGroupChoices()
}
