% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{getRepositories}
\alias{getRepositories}
\title{Get Repositories}
\usage{
getRepositories(network = "", pattern = "", order = TRUE)
}
\arguments{
\item{network}{(character) name of Pandora network}

\item{pattern}{(character) string for meta information search}

\item{order}{(logical) if TRUE order dataframe alphabetically by name}
}
\value{
(data.frame) containing available repositories
}
\description{
Get all vailable repositories or those within
a specific network
optional filtering of meta information for a given string
}
\details{
Formerly getCKANRecordChoices()
}
