% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{getRepositoryList}
\alias{getRepositoryList}
\title{Get Repositories}
\usage{
getRepositoryList(pattern = "", sort = TRUE)
}
\arguments{
\item{pattern}{(character) string for filtering names and entries}

\item{sort}{(logical) if TRUE sort list names alphabetically}
}
\description{
Formerly getCKANRecordChoices()
}
