% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{getResourceList}
\alias{getResourceList}
\title{Get Repositories}
\usage{
getResourceList(fileType = c(), network = "", pattern = "", sort = TRUE)
}
\arguments{
\item{fileType}{(character) list of relevant file types, e.g. c("xls", "xlsx", "csv", "odt")}

\item{network}{(character) name of Pandora network}

\item{pattern}{(character) string for meta information search}

\item{sort}{(logical) if TRUE sort list names alphabetically}
}
\description{
Formerly getCKANResourcesChoices()
}
