% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{getResources}
\alias{getResources}
\title{Get Repositories}
\usage{
getResources(
  fileType = character(),
  repository = "",
  network = "",
  pattern = "",
  order = TRUE
)
}
\arguments{
\item{fileType}{(character) list of relevant file types, e.g. c("xls", "xlsx", "csv", "odt")}

\item{repository}{(character) name of Pandora repository}

\item{network}{(character) name of Pandora network}

\item{pattern}{(character) string for meta information search}

\item{order}{(logical) if TRUE order dataframe alphabetically by name}
}
\value{
(data.frame) containing available resources within a repository
}
\description{
Formerly getCKANResourcesChoices()
}
