% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-getData.R
\name{loadText}
\alias{loadText}
\title{Load Text}
\usage{
loadText(
  path,
  fileEncoding = "",
  collapse = FALSE,
  lineSeparator = "\\n",
  verbose = TRUE
)
}
\arguments{
\item{path}{path or URL to a text file}

\item{fileEncoding}{character string: if non-empty declares the
    encoding used on a file when given as a character string (not on an
    existing connection) so the character data can
    be re-encoded.  See the \sQuote{Encoding} section of the help for
    \code{\link{file}}, {\href{https://cloud.R-project.org/doc/manuals/R-data.html#Variations-on-read_002etable}{\dQuote{Variations on read.table}} in \cite{R Data Import/Export}}, and
    \sQuote{Note}.
  }

\item{collapse}{(logical) if TRUE, collapse all lines to a single string}

\item{lineSeparator}{(character) separator used when collapsing lines}

\item{verbose}{Logical, indicating whether to display processing messages.
If TRUE, messages will be displayed; if FALSE, messages will be suppressed.
Default is TRUE.}
}
\value{
(character vector) lines from the text file, or a single string if
collapse is TRUE
}
\description{
Load Text
}
