% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{renameRepoColumns}
\alias{renameRepoColumns}
\title{Rename Repository Meta Columns}
\usage{
renameRepoColumns(packageList, mapColnames = FALSE)
}
\arguments{
\item{packageList}{(data.frame) optional, output of callAPI() from a previous call to the
Pandora API.}

\item{mapColnames}{(logical) apply names from the 'Additional Info' box from
'https://pandoradata.earth/dataset/' to the columns of returned data}
}
\value{
(data.frame) containing available repositories
}
\description{
Apply names from the 'Additional Info' box from 'https://pandoradata.earth/dataset/' to the
columns of returned data
}
