% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-getData.R
\name{textOptions}
\alias{textOptions}
\title{Text Options}
\usage{
textOptions(sep = ",", dec = ".")
}
\arguments{
\item{sep}{the field separator character.  Values on each line of the
    file are separated by this character.  If \code{sep = ""} (the
    default for \code{read.table}) the separator is \sQuote{white space},
    that is one or more spaces, tabs, newlines or carriage returns.}

\item{dec}{the character used in the file for decimal points.}
}
\value{
a list of extra options for \code{utils::read.csv()}
}
\description{
Text Options
}
