% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-getData.R
\name{validateResource}
\alias{validateResource}
\title{Validate Resource}
\usage{
validateResource(resource)
}
\arguments{
\item{resources}{(data.frame) resources data frame}
}
\value{
(data.frame) resource, or error if empty
}
\description{
Validate Resource
}
