% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-getData.R
\name{xlsxOptions}
\alias{xlsxOptions}
\title{XLSX Options}
\usage{
xlsxOptions(sheet = 1)
}
\arguments{
\item{sheet}{The name or index of the sheet to read data from.}
}
\value{
a list of extra options for \code{openxlsx::read.xlsx()} and \code{readxl::read_excel}
}
\description{
XLSX Options
}
