% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{getFileTypes}
\alias{getFileTypes}
\title{Get File Types}
\usage{
getFileTypes(repository = "", network = "", pattern = "", order = TRUE)
}
\arguments{
\item{repository}{(character) list of relevant file types, e.g. c("xls", "xlsx", "csv", "odt")}

\item{network}{(character) name of Pandora network}

\item{pattern}{(character) string for meta information search}

\item{order}{(logical) if TRUE sort list names alphabetically}
}
\description{
Get File Types
}
